% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_from_labels.R
\name{df_from_labels}
\alias{df_from_labels}
\title{Generate dataframe of variable names and labels from Stata dataset}
\usage{
df_from_labels(path)
}
\arguments{
\item{path}{A path to a Stata dataset.}
}
\value{
dataframe A dataframe with one column for the variable name and one column for the variable label.
}
\description{
This function takes a path to a Stata dataset and returns a dataframe containing
one column for the variable name and one column for the variable label.
}
\examples{
df_from_labels("C:/Users/dougj/Documents/Data/APIS/APIS 2020/APIS 2020 HH.dta")
}
