% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstatr.R
\name{tabstatr}
\alias{tabstatr}
\title{Generate dataframe of weighted mean of several variables.
Placeholder -- change so that the wt argument is optional.}
\usage{
tabstatr(data, vars, wt)
}
\arguments{
\item{data}{Dataframe with columns to calculate mean of}

\item{vars}{<tidy-select> Columns to calculate means of}

\item{wt}{Variable for observation weights}
}
\value{
dataframe
}
\description{
Generate dataframe of weighted mean of several variables.
Placeholder -- change so that the wt argument is optional.
}
\examples{
tabstatr(mtcars, where(is.numeric), wt)
}
